requirements.txt
appnope==0.1.4
asgiref==3.7.2
asttokens==2.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
decorator==5.1.1
Django==4.2.10
django-admin-interface==0.28.5
django-bootstrap-v5==1.0.11
django-bootstrap4==2.2.0
django-colorfield==0.11.0
django-dynamic-breadcrumbs==0.4.3
django-view-breadcrumbs==2.5.1
djangorestframework==3.14.0
executing==2.0.1
ipython==8.12.3
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-slugify==8.0.4
pytz==2024.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
text-unidecode==1.3
traitlets==5.14.1
typing_extensions==4.9.0
wcwidth==0.2.13
